Empty recycle bin when shutting down your Windows
So, as we said, we can choose that the computer itself could simply clean the files as soon as we turn off the computer. But keep in mind that, for this automation, we have to have the tech giant Microsoft’s Windows 10 Pro, as we are going to modify settings that are only available in this version, slightly more complete than the rest.